+++ demo/api/src/config/config.module.ts
@@ -1,14 +1,14 @@
 import { Module } from "@nestjs/common";
 import { ConfigModule as NestConfigModule } from "@nestjs/config";
 import { configNS, EnvironmentVariables } from "@src/config/config.namespace";
-import { plainToClass } from "class-transformer";
+import { plainToInstance } from "class-transformer";
 import { validateSync } from "class-validator";
 
 @Module({
     imports: [
         NestConfigModule.forRoot({
             validate: (config) => {
-                const validatedConfig = plainToClass(EnvironmentVariables, config);
+                const validatedConfig = plainToInstance(EnvironmentVariables, config);
                 const errors = validateSync(validatedConfig, { skipMissingProperties: false });
 
                 if (errors.length > 0) {
@@ -23,7 +23,7 @@ import { validateSync } from "class-validator";
     providers: [
         {
             provide: configNS.KEY,
-            useValue: plainToClass(EnvironmentVariables, configNS()),
+            useValue: plainToInstance(EnvironmentVariables, configNS()),
         },
     ],
     exports: [configNS.KEY],