+++ packages/api/blocks-api/src/blocks/block.ts
@@ -1,6 +1,5 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
-import { classToPlain, plainToClass } from "class-transformer";
-import type { ClassType } from "class-transformer/ClassTransformer";
+import { ClassConstructor, instanceToPlain, plainToInstance } from "class-transformer";
 
 import { createAppliedMigrationsBlockDataFactoryDecorator } from "../migrations/createAppliedMigrationsBlockDataFactoryDecorator";
 import { BlockDataMigrationVersion } from "../migrations/decorators/BlockDataMigrationVersion";
@@ -158,7 +157,7 @@ export abstract class BlockInput<BlockType extends BlockDataInterface = BlockDat
             }),
         );
 
-        return classToPlain(dataWithChildBlocksTransformed) as CreateToPlainReturn<this>;
+        return instanceToPlain(dataWithChildBlocksTransformed) as CreateToPlainReturn<this>;
     }
 }
 
@@ -216,7 +215,7 @@ export type Block<BlockType extends BlockDataInterface = BlockDataInterface, Blo
 const blocks: Block[] = [];
 
 export interface MigrateOptions {
-    migrations: ClassType<BlockMigrationInterface>[];
+    migrations: ClassConstructor<BlockMigrationInterface>[];
     version: number;
 }
 interface CreateBlockOptions {
@@ -227,8 +226,8 @@ interface CreateBlockOptions {
 }
 
 export function createBlock<BlockType extends BlockDataInterface, BlockInputType extends BlockInputInterface>(
-    BlockData: ClassType<BlockType>,
-    BlockInput: ClassType<BlockInputType>,
+    BlockData: ClassConstructor<BlockType>,
+    BlockInput: ClassConstructor<BlockInputType>,
     nameOrOptions: string | CreateBlockOptions,
     {
         // 4th argument is an options-hash
@@ -250,10 +249,10 @@ export function createBlock<BlockType extends BlockDataInterface, BlockInputType
     }
 
     const blockDataFactory: BlockDataFactory<BlockType> = (o) => {
-        return plainToClass(BlockData, o);
+        return plainToInstance(BlockData, o);
     };
 
-    const blockInputFactory: BlockInputFactory<BlockInputType> = (o) => plainToClass(BlockInput, o);
+    const blockInputFactory: BlockInputFactory<BlockInputType> = (o) => plainToInstance(BlockInput, o);
 
     // Decorate BlockDataFactory
     let decorateBlockDataFactory = blockDataFactory;
@@ -314,7 +313,7 @@ export function transformToSave(block: BlockDataInterface): TraversableTransform
     return traverse(block) as TraversableTransformResponse;
 }
 
-export function inputToData<T extends BlockDataInterface, V extends BlockInputInterface>(cls: ClassType<T>, plain: V): T {
+export function inputToData<T extends BlockDataInterface, V extends BlockInputInterface>(cls: ClassConstructor<T>, plain: V): T {
     const dataWithChildBlocksTransformed = Object.fromEntries(
         Object.entries(plain).map(([k, v]) => {
             if (isBlockInputInterface(v)) {
@@ -323,5 +322,5 @@ export function inputToData<T extends BlockDataInterface, V extends BlockInputIn
             return [k, v];
         }),
     );
-    return plainToClass(cls, dataWithChildBlocksTransformed);
+    return plainToInstance(cls, dataWithChildBlocksTransformed);
 }